SetTableCellMargin bug correction
When  tblCellMar element is added to the first position of the tblPr, it goes before tblStyle element. In this case table style settings override cell margin settings and SetTableCellMargin has no visible effect (it was tested with Microsoft Word 2007).
To prevent this behavior, tblCellMar element is added after tblW element. The tblW element is normally added by Table constructor. If tblW is absent, then tblCellMar is added to the beginning of tblPr as it was before (might be better to added another check for tblStyle and add tblCellMar after it if tblW does not exist).